Fix debug-error scenario build: inline enum, import CmuxUpdater in cmuxApp
The "Show Update Error…" submenu names DebugUpdateErrorScenario directly, which
requires cmuxApp to import CmuxUpdater (it only imported CmuxUpdaterUI; the
earlier method-call usage worked through the re-export but a directly-named type
does not). Add the import. Also move the scenario enum + injector into
UpdateStateModel.swift (DEBUG section) rather than a separate file.
